+++ apps/chat-api/src/main.ts
@@ -32,8 +32,8 @@ async function bootstrap() {
             'https://fonts.googleapis.com',
           ],
           fontSrc: ["'self'", 'https://fonts.gstatic.com'],
-          // unpkg.com required for PDF.js worker (pdfjs-dist/build/pdf.worker.min.mjs)
-          scriptSrc: ["'self'", 'https://unpkg.com'],
+          scriptSrc: ["'self'"],
+          workerSrc: ["'self'", 'blob:'],
           imgSrc: ["'self'", 'data:', 'blob:', 'https:'],
         },
       },
