+++ apps/chat/src/hooks/annotations/useAnnotations.ts
@@ -1,6 +1,6 @@
 import type { Annotation, Message } from '@epam/ai-dial-chat-shared';
 import { useMemo } from 'react';
-import { normalizeRawAnnotations } from '../../utils/annotation';
+import { resolveMessageAnnotations } from '../../utils/annotation';
 
 /**
  * Returns the resolved annotation list for an assistant message.
@@ -16,24 +16,8 @@ export const useAnnotations = (
   message: Message,
   isStreaming: boolean,
 ): Annotation[] => {
-  const contentAnnotations = message.custom_content?.annotations;
-  const attachments = message.custom_content?.attachments;
-  const customFields = message['custom_fields'];
-
   return useMemo(() => {
     if (isStreaming) return [];
-
-    if (contentAnnotations?.length) {
-      return contentAnnotations.filter(
-        (a): a is Annotation =>
-          a != null && a.body?.source?.attachment?.url != null,
-      );
-    }
-
-    if (typeof customFields !== 'object' || customFields === null) return [];
-    const raw = (customFields as Record<string, unknown>)['annotations'];
-    if (!Array.isArray(raw) || raw.length === 0) return [];
-
-    return normalizeRawAnnotations(raw, attachments ?? []);
-  }, [isStreaming, contentAnnotations, attachments, customFields]);
+    return resolveMessageAnnotations(message);
+  }, [isStreaming, message]);
 };
