+++ apps/chat/src/hooks/conversation-sources/useConversationSources.ts
@@ -2,6 +2,7 @@ import type { DisplayAttachment, Message } from '@epam/ai-dial-chat-shared';
 import { MessageRole } from '@epam/ai-dial-chat-shared';
 import type { QuotationSource } from '@epam/ai-dial-source-panel';
 import { useMemo } from 'react';
+import { resolveMessageAnnotations } from '../../utils/annotation';
 import { attachmentDtosToDisplayAttachments } from '../../utils/attachment-dto-to-display';
 
 /**
@@ -30,7 +31,7 @@ export const useConversationSources = (
       } else if (msg.role === MessageRole.Assistant) {
         generated.push(...attachments);
 
-        for (const annotation of msg.custom_content?.annotations ?? []) {
+        for (const annotation of resolveMessageAnnotations(msg)) {
           const att = annotation?.body?.source?.attachment;
           const url = att?.url;
           if (!url || seenUrls.has(url)) continue;
